Moved helpers to `helperFunctions`
