+++ src/utilities/helperFunctions.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export const validate = (input, setErrorMessage) => {
   let validated = false;
   // A city name must include at least one unicode letter, and may include numbers and spaces
@@ -13,4 +12,38 @@ export const validate = (input, setErrorMessage) => {
   }
 
   return validated;
+};
+
+export const toCapitalized = function capitalizeString(str) {
+  if (!str) return '';
+
+  const words = str.split(' ');
+  const capitalizedWords = words.map(
+    (word) => `${word.charAt(0).toUpperCase()}${word.substring(1)}`
+  );
+
+  return capitalizedWords.join(' ');
+};
+
+export const toLocalDate = function convertTimestampToLocalDate(timestamp) {
+  return new Date(timestamp * 1000);
+};
+
+export const getHourString = function getFormattedHourStringFromDate(date) {
+  const hoursIn24Format = date.getHours();
+  let hourString;
+
+  if (hoursIn24Format < 12) {
+    if (hoursIn24Format === 0) hourString = 'midnight';
+    else hourString = `${hoursIn24Format} am`;
+  } else if (hoursIn24Format > 12) {
+    if (hoursIn24Format === 12) hourString = 'noon';
+    else hourString = `${hoursIn24Format - 12} pm`;
+  }
+
+  return hourString;
+};
+
+export const getIconSrc = function getWeatherIconSrcURL(iconId) {
+  return `https://openweathermap.org/img/wn/${iconId}@2x.png`;
 };
